wip(#1101): extract lower/template
